increase HT requirement in confSL.py

+++ ntupler/python/confSL.py
@@ -34,7 +34,7 @@ process.tree.is2016 = cms.bool(True)
 process.tree.deBug = cms.bool(False)
 process.tree.verBose = cms.bool(False)
 process.tree.maskDeepCSV = cms.bool(False)
-process.tree.HTMin = cms.double(50)
+process.tree.HTMin = cms.double(500)
 process.tree.NjMin = cms.int32(3)
 
 process.TFileService = cms.Service("TFileService",
